Add files via upload
The first App I developed. We worked in an agile flow over 2 sprints in a team of 8. This is a small account documenting our progress with a walk through of the app in its 2nd in its 2nd prototype phase. I worked on the functionality side of the app developing flows and researching what was needed and what was available in the market.